Deal with default envadmin and email address
If we have the email address, we want to set the environment admin to that name, even if that means overwriting the environment admin value that was passed in.
If we don't, we set the environment admin to the value passed in.
Otherwise we default to the admin name of "user"